pyproject.toml
[project]
name = "nonebot-plugin-uninfo"
version = "0.1.0"
description = "Universal Information Model for Nonebot2"
authors = [
    {name = "RF-Tar-Railt", email = "rf_tar_railt@qq.com"},
]
dependencies = [
    "nonebot2>=2.3.0",
    "importlib-metadata>=4.13.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "isort>=5.13.2",
    "black>=24.4.2",
    "loguru>=0.7.2",
    "ruff>=0.4.2",
    "fastapi>=0.110.2",
    "uvicorn[standard]>=0.28.1",
    "pydantic>=2.7.0",
    "nonebot2[httpx,websockets]>=2.3.0",
    "nonebot-adapter-console>=0.5.0",
    "nonebot-adapter-ding>=2.0.0a16",
    "nonebot-adapter-discord>=0.1.8",
    "nonebot-adapter-dodo>=0.2.1",
    "nonebot-adapter-feishu>=2.6.0",
    "nonebot-adapter-kaiheila>=0.3.4",
    "nonebot-adapter-mirai>=2.3.2",
    "nonebot-adapter-mirai2>=0.0.22",
    "nonebot-adapter-minecraft>=1.2.2",
    "nonebot-adapter-onebot>=2.4.3",
    "nonebot-adapter-qq>=1.4.4",
    "nonebot-adapter-red>=0.9.0",
    "nonebot-adapter-satori>=0.12.3",
    "nonebot-adapter-telegram>=0.1.0b17",
    "nonebot-adapter-kritor>=0.1.1",
    "nonebot-adapter-tailchat>=0.1.0b7",
]
[tool.pdm.build]
includes = ["src/nonebot_plugin_uninfo"]


[tool.black]
line-length = 120
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 120
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["C901", "T201", "E731", "E402"]

[tool.pyright]
pythonVersion = "3.9"
pythonPlatform = "All"
defineConstant = { PYDANTIC_V2 = true }
typeCheckingMode = "basic"
reportShadowedImports = false
disableBytesTypePromotions = true

[tool.pdm.scripts]
format = { composite = ["isort ./src/ ","black ./src/ ","ruff check ./src/"] }


[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["example"]
plugin_dirs = []
builtin_plugins = []
